리팩토링 시도
